+++ Locale/MRP_Locale.ruRU.lua
@@ -19,6 +19,7 @@ MRP.L = {
   [" - /mrp tomtom on|off → (Deprecated) Use /mrp waypoint instead"] = " - /mrp tomtom on|off → (Устарело) Используйте /mrp waypoint",
   [" - /mrp update → Show official download sources"] = " - /mrp update → Show official download sources", -- MISSING TRANSLATION / SAME AS ENGLISH,
   [" - /mrp updatedisplaydelayed <number> → Force update of the display after a delay (default: 0.25 seconds)"] = " - /mrp updatedisplaydelayed <число> → Принудительное обновление экрана с задержкой (по умолчанию 0.25 сек)",
+  [" - /mrp versions → Print addon and library versions"] = " - /mrp versions → Print addon and library versions", -- MISSING TRANSLATION / SAME AS ENGLISH,
   [" - /mrp waypoint none|waypoint|tomtom → Choose the waypoint system"] = " - /mrp waypoint none|waypoint|tomtom → Выбор системы путевых точек",
   ["*Normal difficulty dungeons can be run up to 10 times per hour"] = "*Подземелья обычной сложности можно проходить до 10 раз в час",
   ["%s: This boss is dead"] = "%s: Босс убит",